added module with class BWT and its documentation; replaced array with dictionary; fixed some details.
