--- Burrows-Wheeler/Burrows-Wheeler.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34024.191
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Burrows-Wheeler", "Burrows-Wheeler.csproj", "{D0DCFCDB-0873-4813-9E8B-4E2B6784295C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D0DCFCDB-0873-4813-9E8B-4E2B6784295C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D0DCFCDB-0873-4813-9E8B-4E2B6784295C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D0DCFCDB-0873-4813-9E8B-4E2B6784295C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D0DCFCDB-0873-4813-9E8B-4E2B6784295C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {189AD0B0-632C-4461-9DE2-8139519A720B}
-	EndGlobalSection
-EndGlobal
