+++ BurrowsWheeler/BurrowsWheeler.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34622.214
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BurrowsWheeler", "BurrowsWheeler\BurrowsWheeler.csproj", "{433FDB61-5744-402A-98E9-04EDD67BCCD7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{433FDB61-5744-402A-98E9-04EDD67BCCD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{433FDB61-5744-402A-98E9-04EDD67BCCD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{433FDB61-5744-402A-98E9-04EDD67BCCD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{433FDB61-5744-402A-98E9-04EDD67BCCD7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {18C5D986-5794-4454-AC97-4A71B7E95DD3}
+	EndGlobalSection
+EndGlobal
